Edit test cases
