+++ src/web/app/pages-session/session-result-page/session-result-page.component.spec.ts
@@ -11,19 +11,12 @@ import { LogService } from '../../../services/log.service';
 import { NavigationService } from '../../../services/navigation.service';
 import { StudentService } from '../../../services/student.service';
 import {
-  AuthInfo, FeedbackContributionQuestionDetails, FeedbackContributionResponseDetails,
-  FeedbackMcqQuestionDetails, FeedbackMcqResponseDetails,
-  FeedbackParticipantType,
-  FeedbackQuestion,
-  FeedbackQuestionType, FeedbackRubricQuestionDetails, FeedbackRubricResponseDetails,
+  AuthInfo,
   FeedbackSession,
   FeedbackSessionPublishStatus,
-  FeedbackSessionSubmissionStatus, FeedbackTextQuestionDetails, FeedbackTextResponseDetails,
-  FeedbackVisibilityType,
-  NumberOfEntitiesToGiveFeedbackToSetting,
-  QuestionOutput,
+  FeedbackSessionSubmissionStatus,
   RegkeyValidity,
-  ResponseVisibleSetting, SessionResults,
+  ResponseVisibleSetting,
   SessionVisibleSetting,
 } from '../../../types/api-output';
 import { Intent } from '../../../types/api-request';
@@ -66,123 +59,6 @@ describe('SessionResultPageComponent', () => {
       isStudent: false,
       isMaintainer: false,
     },
-  };
-
-  const testQuestion1: FeedbackQuestion = {
-    feedbackQuestionId: 'feedbackQuestion1',
-    questionNumber: 1,
-    questionBrief: 'How well did team member perform?',
-    questionDescription: '',
-    questionDetails: {
-      hasAssignedWeights: false,
-      mcqWeights: [],
-      mcqOtherWeight: 0,
-      mcqChoices: [
-        '<p>Good</p>',
-        '<p>Normal</p>',
-        '<p>Bad</p>',
-      ],
-      otherEnabled: false,
-      generateOptionsFor: 'NONE',
-      questionType: FeedbackQuestionType.MCQ,
-      questionText: 'How well did team member perform?',
-    } as FeedbackMcqQuestionDetails,
-    questionType: FeedbackQuestionType.MCQ,
-    giverType: FeedbackParticipantType.STUDENTS,
-    recipientType: FeedbackParticipantType.OWN_TEAM_MEMBERS_INCLUDING_SELF,
-    numberOfEntitiesToGiveFeedbackToSetting: NumberOfEntitiesToGiveFeedbackToSetting.UNLIMITED,
-    showResponsesTo: [],
-    showGiverNameTo: [],
-    showRecipientNameTo: [],
-    customNumberOfEntitiesToGiveFeedbackTo: 0,
-  };
-
-  const testQuestion2: FeedbackQuestion = {
-    feedbackQuestionId: 'feedbackQuestion2',
-    questionNumber: 2,
-    questionBrief: 'Rate your teammates in contribution',
-    questionDescription: '',
-    questionDetails: {
-      questionType: FeedbackQuestionType.CONTRIB,
-      questionText: 'Rate your teammates in contribution',
-      isNotSureAllowed: false,
-    } as FeedbackContributionQuestionDetails,
-    questionType: FeedbackQuestionType.CONTRIB,
-    giverType: FeedbackParticipantType.STUDENTS,
-    recipientType: FeedbackParticipantType.OWN_TEAM_MEMBERS,
-    numberOfEntitiesToGiveFeedbackToSetting: NumberOfEntitiesToGiveFeedbackToSetting.UNLIMITED,
-    showResponsesTo: [],
-    showGiverNameTo: [],
-    showRecipientNameTo: [],
-    customNumberOfEntitiesToGiveFeedbackTo: 0,
-  };
-
-  const testQuestion3: FeedbackQuestion = {
-    feedbackQuestionId: 'feedbackQuestion3',
-    questionNumber: 3,
-    questionBrief: 'Rate your teammates proficiency',
-    questionDescription: '',
-    questionDetails: {
-      questionType: FeedbackQuestionType.RUBRIC,
-      questionText: 'Rate your teammates proficiency',
-      hasAssignedWeights: false,
-      rubricWeightsForEachCell: [[]],
-      rubricChoices: ['Poor', 'Average', 'Good'],
-      rubricSubQuestions: [],
-      rubricDescriptions: [[]],
-    } as FeedbackRubricQuestionDetails,
-    questionType: FeedbackQuestionType.RUBRIC,
-    giverType: FeedbackParticipantType.STUDENTS,
-    recipientType: FeedbackParticipantType.OWN_TEAM_MEMBERS,
-    numberOfEntitiesToGiveFeedbackToSetting: NumberOfEntitiesToGiveFeedbackToSetting.UNLIMITED,
-    showResponsesTo: [],
-    showGiverNameTo: [],
-    showRecipientNameTo: [],
-    customNumberOfEntitiesToGiveFeedbackTo: 0,
-  };
-
-  const testQuestionAnonymousResponse1: FeedbackQuestion = {
-    feedbackQuestionId: 'feedbackQuestionAnonymousResponse1',
-    questionNumber: 1,
-    questionBrief: 'What comments do you have regarding each of your team members? '
-        + '(response is confidential and will only be shown to the instructor).',
-    questionDescription: '',
-    questionDetails: {
-      shouldAllowRichText: true,
-      questionType: FeedbackQuestionType.TEXT,
-      questionText: 'What comments do you have regarding each of your team members? '
-          + '(response is confidential and will only be shown to the instructor).',
-    } as FeedbackTextQuestionDetails,
-    questionType: FeedbackQuestionType.TEXT,
-    giverType: FeedbackParticipantType.STUDENTS,
-    recipientType: FeedbackParticipantType.OWN_TEAM_MEMBERS,
-    numberOfEntitiesToGiveFeedbackToSetting: NumberOfEntitiesToGiveFeedbackToSetting.UNLIMITED,
-    showResponsesTo: [FeedbackVisibilityType.INSTRUCTORS],
-    showGiverNameTo: [FeedbackVisibilityType.INSTRUCTORS],
-    showRecipientNameTo: [FeedbackVisibilityType.INSTRUCTORS],
-    customNumberOfEntitiesToGiveFeedbackTo: 0,
-  };
-
-  const testQuestionAnonymousResponse2: FeedbackQuestion = {
-    feedbackQuestionId: 'feedbackQuestionAnonymousResponse2',
-    questionNumber: 2,
-    questionBrief: 'How are the team dynamics thus far? '
-        + '(response is confidential and will only be shown to the instructor).',
-    questionDescription: '',
-    questionDetails: {
-      shouldAllowRichText: true,
-      questionType: FeedbackQuestionType.TEXT,
-      questionText: 'How are the team dynamics thus far? '
-          + '(response is confidential and will only be shown to the instructor).',
-    } as FeedbackTextQuestionDetails,
-    questionType: FeedbackQuestionType.TEXT,
-    giverType: FeedbackParticipantType.STUDENTS,
-    recipientType: FeedbackParticipantType.OWN_TEAM,
-    numberOfEntitiesToGiveFeedbackToSetting: NumberOfEntitiesToGiveFeedbackToSetting.UNLIMITED,
-    showResponsesTo: [FeedbackVisibilityType.INSTRUCTORS],
-    showGiverNameTo: [FeedbackVisibilityType.INSTRUCTORS],
-    showRecipientNameTo: [FeedbackVisibilityType.INSTRUCTORS],
-    customNumberOfEntitiesToGiveFeedbackTo: 0,
   };
 
   let component: SessionResultPageComponent;
@@ -325,267 +201,36 @@ describe('SessionResultPageComponent', () => {
     expect(fixture).toMatchSnapshot();
   });
 
-  it('should snap with feedback session with questions', () => {
-    component.session = testFeedbackSession;
-    component.questions = [
-      {
-        feedbackQuestion: testQuestion1,
-        questionStatistics: '',
-        allResponses: [],
-        responsesToSelf: [],
-        responsesFromSelf: [
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-1',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'recipient1',
-            recipientTeam: 'team1',
-            recipientSection: 'section1',
-            responseDetails: {
-              answer: 'Good',
-              isOther: false,
-              otherFieldContent: '',
-            } as FeedbackMcqResponseDetails,
-            instructorComments: [],
-          },
-        ],
-        otherResponses: [[]],
-      },
-      {
-        feedbackQuestion: testQuestion2,
-        questionStatistics: '',
-        allResponses: [],
-        responsesToSelf: [
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-2',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'giver1',
-            recipientTeam: 'team1',
-            recipientSection: 'section1',
-            responseDetails: {
-              answer: 120,
-            } as FeedbackContributionResponseDetails,
-            instructorComments: [],
-          },
-        ],
-        responsesFromSelf: [
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-3',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'recipient2',
-            recipientTeam: 'team2',
-            recipientSection: 'section2',
-            responseDetails: {
-              answer: 110,
-            } as FeedbackContributionResponseDetails,
-            instructorComments: [],
-          },
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-4',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'recipient3',
-            recipientTeam: 'team2',
-            recipientSection: 'section2',
-            responseDetails: {
-              answer: 100,
-            } as FeedbackContributionResponseDetails,
-            instructorComments: [],
-          },
-        ],
-        otherResponses: [[]],
-      },
-      {
-        feedbackQuestion: testQuestion3,
-        questionStatistics: '',
-        allResponses: [],
-        responsesToSelf: [],
-        responsesFromSelf: [
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-5',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'recipient3',
-            recipientTeam: 'team2',
-            recipientSection: 'section2',
-            responseDetails: {
-              answer: [1],
-            } as FeedbackRubricResponseDetails,
-            instructorComments: [
-              {
-                commentGiver: 'comment-giver-1',
-                lastEditorEmail: 'comment@egeg.com',
-                feedbackResponseCommentId: 1,
-                commentText: 'this is a text',
-                createdAt: 1402775804,
-                lastEditedAt: 1402775804,
-                isVisibilityFollowingFeedbackQuestion: true,
-                showGiverNameTo: [],
-                showCommentTo: [],
-              },
-            ],
-          },
-        ],
-        otherResponses: [[]],
-      },
-    ];
+  // it('should load feedback questions', () => {
+  //   const testFeedbackQuestions: FeedbackQuestions = {
+  //     questions: [
+  //       {
+  //         feedbackQuestionId: testMcqQuestionSubmissionForm2.feedbackQuestionId,
+  //         questionNumber: testMcqQuestionSubmissionForm2.questionNumber,
+  //         questionBrief: testMcqQuestionSubmissionForm2.questionBrief,
+  //         questionDescription: testMcqQuestionSubmissionForm2.questionDescription,
+  //         questionDetails: testMcqQuestionSubmissionForm2.questionDetails,
+  //         questionType: testMcqQuestionSubmissionForm2.questionType,
+  //         giverType: testMcqQuestionSubmissionForm2.giverType,
+  //         recipientType: testMcqQuestionSubmissionForm2.recipientType,
+  //         numberOfEntitiesToGiveFeedbackToSetting:
+  //           testMcqQuestionSubmissionForm2.numberOfEntitiesToGiveFeedbackToSetting,
+  //         customNumberOfEntitiesToGiveFeedbackTo: testMcqQuestionSubmissionForm2.customNumberOfEntitiesToGiveFeedbackTo,
+  //         showResponsesTo: testMcqQuestionSubmissionForm2.showResponsesTo,
+  //         showGiverNameTo: testMcqQuestionSubmissionForm2.showGiverNameTo,
+  //         showRecipientNameTo: testMcqQuestionSubmissionForm2.showRecipientNameTo,
+  //       },
+  //     ],
+  //   };
 
-    fixture.detectChanges();
-    expect(fixture).toMatchSnapshot();
-  });
+  //   const getQuestionsSpy: SpyInstance = jest.spyOn(feedbackQuestionsService, 'getFeedbackQuestions')
+  //       .mockReturnValue(of(testFeedbackQuestions));
 
-  it('should snap with feedback session with questions of anonymous responses', () => {
-    component.session = testFeedbackSession;
-    component.questions = [
-      {
-        feedbackQuestion: testQuestionAnonymousResponse1,
-        questionStatistics: '',
-        allResponses: [
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-1',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'recipient1',
-            recipientTeam: 'team1',
-            recipientSection: 'section1',
-            responseDetails: {
-              answer: '<p>asdf</p>',
-            } as FeedbackTextResponseDetails,
-            instructorComments: [],
-          },
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-2',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'recipient2',
-            recipientTeam: 'team1',
-            recipientSection: 'section1',
-            responseDetails: {
-              answer: '<p>asdf</p>',
-            } as FeedbackTextResponseDetails,
-            instructorComments: [],
-          },
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-3',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'recipient3',
-            recipientTeam: 'team1',
-            recipientSection: 'section1',
-            responseDetails: {
-              answer: '<p>asdf</p>',
-            } as FeedbackTextResponseDetails,
-            instructorComments: [],
-          },
-        ],
-        responsesToSelf: [],
-        responsesFromSelf: [
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-1',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'recipient1',
-            recipientTeam: 'team1',
-            recipientSection: 'section1',
-            responseDetails: {
-              answer: '<p>asdf</p>',
-            } as FeedbackTextResponseDetails,
-            instructorComments: [],
-          },
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-2',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'recipient2',
-            recipientTeam: 'team1',
-            recipientSection: 'section1',
-            responseDetails: {
-              answer: '<p>asdf</p>',
-            } as FeedbackTextResponseDetails,
-            instructorComments: [],
-          },
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-3',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'recipient3',
-            recipientTeam: 'team1',
-            recipientSection: 'section1',
-            responseDetails: {
-              answer: '<p>asdf</p>',
-            } as FeedbackTextResponseDetails,
-            instructorComments: [],
-          },
-        ],
-        otherResponses: [],
-      },
-      {
-        feedbackQuestion: testQuestionAnonymousResponse2,
-        questionStatistics: '',
-        allResponses: [
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-4',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'team1',
-            recipientTeam: 'team1',
-            recipientSection: 'section1',
-            responseDetails: {
-              answer: '<p>asdf</p>',
-            } as FeedbackTextResponseDetails,
-            instructorComments: [],
-          },
-        ],
-        responsesToSelf: [],
-        responsesFromSelf: [
-          {
-            isMissingResponse: false,
-            responseId: 'resp-id-4',
-            giver: 'giver1',
-            giverTeam: 'team1',
-            giverSection: 'section1',
-            recipient: 'team1',
-            recipientTeam: 'team1',
-            recipientSection: 'section1',
-            responseDetails: {
-              answer: '<p>asdf</p>',
-            } as FeedbackTextResponseDetails,
-            instructorComments: [],
-          },
-        ],
-        otherResponses: [],
-      },
-    ];
-
-    fixture.detectChanges();
-    expect(fixture).toMatchSnapshot();
-  });
+  //   expect(getQuestionsSpy).toHaveBeenLastCalledWith(getFeedbackSessionArgs);
+  //   expect(component.questionSubmissionForms.length).toEqual(1);
+  //   expect(component.questionSubmissionForms[0]).toEqual(testMcqQuestionSubmissionForm2);
+  //   expect(component.questionsNeedingSubmission.length).toEqual(0);
+  // });
 
   it('should fetch auth info on init', () => {
     jest.spyOn(authService, 'getAuthUser').mockReturnValue(of(testInfo));
@@ -638,61 +283,6 @@ describe('SessionResultPageComponent', () => {
     expect(component.personName).toEqual('student-name');
     expect(component.session.courseId).toEqual('CS1231');
     expect(logSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('should fetch session results when loading feedback session', () => {
-    const testValidity: RegkeyValidity = {
-      isAllowedAccess: true,
-      isUsed: false,
-      isValid: false,
-    };
-
-    const testFeedbackSessionResult: SessionResults = {
-      questions: [
-        {
-          feedbackQuestion: testQuestion1,
-          questionStatistics: '',
-          allResponses: [],
-          responsesToSelf: [],
-          responsesFromSelf: [],
-          otherResponses: [],
-        },
-        {
-          feedbackQuestion: testQuestion3,
-          questionStatistics: '',
-          allResponses: [],
-          responsesToSelf: [],
-          responsesFromSelf: [],
-          otherResponses: [],
-        },
-        {
-          feedbackQuestion: testQuestion2,
-          questionStatistics: '',
-          allResponses: [],
-          responsesToSelf: [],
-          responsesFromSelf: [],
-          otherResponses: [],
-        },
-      ],
-    };
-
-    jest.spyOn(authService, 'getAuthUser').mockReturnValue(of(testInfo));
-    jest.spyOn(authService, 'getAuthRegkeyValidity').mockReturnValue(of(testValidity));
-    jest.spyOn(feedbackSessionService, 'getFeedbackSession').mockReturnValue(of(testFeedbackSession));
-    const fsSpy: SpyInstance = jest.spyOn(feedbackSessionService, 'getFeedbackSessionResults')
-        .mockReturnValue(of(testFeedbackSessionResult));
-
-    component.ngOnInit();
-
-    expect(fsSpy).toHaveBeenCalledTimes(1);
-    expect(fsSpy).toHaveBeenLastCalledWith({
-      courseId: 'CS3281',
-      feedbackSessionName: 'Peer Feedback',
-      intent: Intent.STUDENT_RESULT,
-      key: 'reg-key',
-    });
-    expect(component.questions.map((question: QuestionOutput) => question.feedbackQuestion.questionNumber))
-        .toEqual([1, 2, 3]);
   });
 
   it('should deny access for reg key not belonging to logged in user', () => {
